Refactored how key presses are handled
Moved the logic for dealing with key presses into its own function which
sets the values of the passed key_press_t. This way, the key_press_t can
be passed to update functions.

Added an update function for screen which will react to the arrow key
presses. It's not completely done as I want to add more functionality to
it but it's good enough for now.